requirements.txt
Django>=3.0.7,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2-binary>=2.8.5,<2.9.0
django-nose>=1.4.6,<1.5.0
coverage>=5.2.1,<5.3.0
factory-boy>=3.0.1,<3.1.0
Faker>=4.1.1,<4.2.0

flake8>=3.8.3,<3.9.0